+++ WebApp/proxy.config.json
@@ -1,5 +1,5 @@
 {
-    "/backend/*":{
+    "/api/*":{
         "target":   "http://localhost:8080",
         "secure":   false,
         "Loglevel": "debug",
