+++ WebApp/src/app/common-menu/common-menu.component.scss
@@ -11,7 +11,7 @@ mat-sidenav {
     color: white;
     border-radius: 10px;
     padding: 16px;
-    text-align: center;
+    text-align: left;
 }
 
 .side-menu-icon{
